+++ frontend/src/components/restaurant/FavoriteContent.tsx
@@ -3,8 +3,6 @@ import {
     Button,
     ButtonGroup,
     Flex,
-    Grid,
-    GridItem,
     Heading,
     Image,
     Popover,
@@ -17,6 +15,7 @@ import {
     PopoverTrigger,
     Show,
     Text,
+    useDisclosure,
 } from "@chakra-ui/react"
 import React, { FC, useState } from "react"
 import { AiOutlineClose } from "react-icons/ai"
@@ -33,12 +32,7 @@ const FavoriteContent: FC<{
     link: string
     img: string
 }> = ({ id,resName, phone, open, close, website,link, img }) => {
-    // const ClosePop = () => {
-    //     setclose(true);
-    //     // console.log(close);
-
-    // }
-
+    const {onClose} = useDisclosure()
     const DeleteFv = () => {
         Restaurant[id].status = false
         console.log(Restaurant[id].status)
@@ -46,19 +40,12 @@ const FavoriteContent: FC<{
     
 
     return (
-        // <Box mb={{lg:"4rem"}} width="100%">
-        //    {/* <Flex direction={"row"} justifyContent={"center"}>
-        //     <Show above="lg">
-        //             <Box width={"30%"} mt={"35px"}>
-        //                 <Image boxSize="12.5rem" height={"9.6rem"} src="https://bit.ly/dan-abramov" alt="Dan Abramov" borderRadius={"10px"} />
-        //             </Box>
-        //         </Show> */}
         <>
             <Show below="sm">
-                <Box width={"100%"} mt={"25px"} backgroundColor={"white"} p={"5"} borderRadius="lg" boxShadow={"lg"}>
+            <Box width={"100%"} mt={"25px"} backgroundColor={"white"} p={"5"} borderRadius="lg" boxShadow={"lg"} >
                     <Popover placement="auto">
-                        {({ isOpen, onClose }) => (
-                            <>
+                        {/* {({onClose}) => (
+                            <> */}
                                 <PopoverTrigger>
                                     <Flex justifyContent={"end"} height={3}>
                                         <AiOutlineClose />
@@ -81,24 +68,17 @@ const FavoriteContent: FC<{
                                         </ButtonGroup>
                                     </PopoverFooter>
                                 </PopoverContent>
-                            </>
-                        )}
+                            {/* </>
+                        )} */}
                     </Popover>
 
-                    <Flex mb={"15px"}>
-                        {/* <Show below="lg"> */}
-                        <Box>
-                            <Image
-                                boxSize={{ base: "12rem", lg: "12.5rem" }}
-                                height={{base:"5.5rem", lg: "9.5rem" }}
-                                src={img}
-                                alt="Dan Abramov"
-                                borderRadius={"10px"}
-                            />
+                    <Flex alignItems={"center"}>
+                    <Box width={"30%"} ml={"1rem"}>
+                            <Image boxSize="5rem" src={img} alt="Dan Abramov" borderRadius={"10px"} />
                         </Box>
                         {/* </Show> */}
-                        <Box width={"100%"} display="flex" justifyContent={"center"}>
-                            <Box width={"70%"} color={"black"}>
+                       
+                        <Box width={"60%"} color={"black"}>
                                 <Text fontSize={{ base: "sm", lg: "lg" }}>
                                     <span style={{ fontWeight: "bold" }}>Name:</span> {resName}
                                 </Text>
@@ -112,14 +92,14 @@ const FavoriteContent: FC<{
                                     <span style={{ fontWeight: "bold" }}>Website:</span> <a href={website}>{resName}</a>
                                 </Text>
                             </Box>
-                        </Box>
+                       
                     </Flex>
                 </Box>
             </Show>
             <Show above="sm">
                 <Box width={"100%"} p={5} shadow="md" borderWidth="1px" mt={"35px"} borderRadius={"lg"}>
                     <Popover placement="bottom">
-                        {({ isOpen, onClose }) => (
+                        {({ onClose }) => (
                             <>
                                 <PopoverTrigger>
                                     <Flex justifyContent={"end"} height={3}>
@@ -153,7 +133,6 @@ const FavoriteContent: FC<{
                             </Heading>
                             <Image
                                 boxSize={"12.5rem"}
-                                // height={{ lg: "9.5rem" }}
                                 src={img}
                                 alt="Dan Abramov"
                                 borderRadius={"10px"}
